Student_Record_Management_System.c
This project is a simple student record management system implemented in C. It allows users to perform various operations such as adding a new student record, displaying a student record, deleting a student record, filling in the marks of a student record, searching for a student, and displaying a report card of a specific student. The program uses a struct to store student information such as ID, name, birth date, address, class, emergency contact, and average. It also uses a 2D array to store student marks for different subjects. The system provides a user-friendly menu interface for easy navigation.